% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny.r
\name{renderStreamgraph}
\alias{renderStreamgraph}
\title{Widget render function for use in Shiny}
\usage{
renderStreamgraph(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

