% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/marker.r
\name{sg_add_marker}
\alias{sg_add_marker}
\title{Add a vertical marker (with optional label) to streamgraph}
\usage{
sg_add_marker(sg, x, label = "", stroke_width = 0.5, stroke = "#7f7f7f",
  space = 5, y = 0, color = "#7f7f7f", size = 12, anchor = "start")
}
\arguments{
\item{x}{horizontal position}

\item{label}{text for the annotation}

\item{stroke_width}{line width}

\item{stroke}{line color}

\item{space}{space (in points) from the marker to place the label}

\item{y}{vertical position}

\item{color}{color of the label}

\item{size}{font size#'}

\item{anchor}{how to justify the label (one of \code{start} [left],
\code{middle} [center] or \code{end} [right])}
}
\description{
This is useful for marking/labeling notable events along the streams.
}

