% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{sg_annotate}
\alias{sg_annotate}
\title{Add text annotation to streamgraph}
\usage{
sg_annotate(sg, label, x, y, color = "black", size = 12)
}
\arguments{
\item{label}{text for the annotation}

\item{x}{horizontal position}

\item{y}{vertical position}

\item{color}{color of the label}

\item{size}{font size}
}
\description{
Use this function to place text at any point on a streamgraph. This
is especially useful for non-interactive streamgraphs (i.e. to label
a particular stream).
}

