% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg_axis_x.r
\name{sg_axis_x}
\alias{sg_axis_x}
\title{Modify streamgraph x axis formatting}
\usage{
sg_axis_x(sg, tick_interval = NULL, tick_units = NULL, tick_format = NULL)
}
\arguments{
\item{sg}{streamgraph object}

\item{tick_interval}{tick interval}

\item{tick_units}{units for the ticks}

\item{tick_format}{how to show the labels (subset of \code{strftime}
formatters for \code{date} scale, otherwise \code{sprintf} formats for
\code{continuous} scale) (defaults to \code{\%b} - must specify if \code{continuous}).
See \href{D3 formatting}{https://github.com/mbostock/d3/wiki/Formatting} for more details.}
}
\value{
streamgraph object
}
\description{
Change the tick interval, units and label text display format for the
streamgraph x axis.
}
\examples{
\dontrun{
library(dplyr)
library(streamgraph)
ggplot2movies::movies \%>\%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) \%>\%
  tidyr::gather(genre, value, -year) \%>\%
  group_by(year, genre) \%>\%
  tally(wt=value) \%>\%
  ungroup \%>\%
  mutate(year=as.Date(sprintf("\%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year") \%>\%
  sg_axis_x(20, "year", "\%Y")
}
}

