% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{sg_colors}
\alias{sg_colors}
\title{Modify streamgraph colors}
\usage{
sg_colors(sg, palette = NULL, axis_color = "black",
  tooltip_color = "black", label_color = "black")
}
\arguments{
\item{sg}{streamgraph object}

\item{palette}{UNUSED; being removed in next release; use \code{sg_fill_*} instead}

\item{axis_color}{color of the axis text (defaults to "\code{black}")}

\item{tooltip_color}{color of the tooltip text (defaults to "\code{black}")}

\item{label_color}{color of the label text for the legend select menu (defaults to "\code{black}")}
}
\value{
streamgraph object
}
\description{
Change the ColorBrewer palette being used
}
\examples{
\dontrun{
library(dplyr)
library(streamgraph)
ggplot2::movies \%>\%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) \%>\%
  tidyr::gather(genre, value, -year) \%>\%
  group_by(year, genre) \%>\%
  tally(wt=value) \%>\%
  ungroup \%>\%
  mutate(year=as.Date(sprintf("\%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year") \%>\%
  sg_colors("PuOr")
}
}

