% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{sg_fill_brewer}
\alias{sg_fill_brewer}
\title{Use ColorBrewer palettes for streamgraph fill colors}
\usage{
sg_fill_brewer(sg, palette = "Spectral")
}
\arguments{
\item{sg}{streamgraph object}

\item{palette}{ColorBrewer pallete atomic character value (defaults to \code{Spectral})}
}
\value{
streamgraph object
}
\description{
ColorBrewer provides sequential, diverging and qualitative colour schemes
which are particularly suited and tested to display categorical values.
}
\examples{
\dontrun{
library(dplyr)
library(streamgraph)
ggplot2movies::movies \%>\%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) \%>\%
  tidyr::gather(genre, value, -year) \%>\%
  group_by(year, genre) \%>\%
  tally(wt=value) \%>\%
  ungroup \%>\%
  mutate(year=as.Date(sprintf("\%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year") \%>\%
  sg_fill_brewer("PuOr")
}
}
