% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/color.r
\name{sg_fill_manual}
\alias{sg_fill_manual}
\title{Use manual colors for streamgraph fill colors}
\usage{
sg_fill_manual(sg, values = NULL)
}
\arguments{
\item{sg}{streamgraph object}

\item{values}{character vector of}
}
\value{
streamgraph object
}
\description{
Specify a vector of colors (e.g. \code{c("red", "#00ff00", rgb(0,0,1))}) to
use for the color scale. Note that \code{streamgraph} sorts the categorical values
before assigning the mappings, which means you can use that as a determinstic way of
assigning specific colors to categories. If the number of categories
exceeds the number of colors in the palette, the colors will be reused in order.
}
\examples{
\dontrun{
library(dplyr)
library(streamgraph)
ggplot2::movies \%>\%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) \%>\%
  tidyr::gather(genre, value, -year) \%>\%
  group_by(year, genre) \%>\%
  tally(wt=value) \%>\%
  ungroup \%>\%
  mutate(year=as.Date(sprintf("\%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year") \%>\%
  sg_fill_manual(c("black", "#ffa500", "blue", "white", "#00ff00', "red"))
}
}

