% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{sg_fill_tableau}
\alias{sg_fill_tableau}
\title{Use Tableau discrete palettes for streamgraph fill colors}
\usage{
sg_fill_tableau(sg, palette = "tableau20")
}
\arguments{
\item{sg}{streamgraph object}

\item{palette}{Tableau discrete pallete atomic character value (defaults to \code{tableau20}). Must be one of
\code{c("tableau20", "tableau10medium", "gray5", "colorblind10", "trafficlight", "purplegray12", "bluered12", "greenorange12", "cyclic")}}
}
\value{
streamgraph object
}
\description{
Tableau discrete palettes provide colour schemes
which are particularly suited and tested to display categorical values.
}
\examples{
\dontrun{
library(dplyr)
library(streamgraph)
ggplot2::movies \%>\%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) \%>\%
  tidyr::gather(genre, value, -year) \%>\%
  group_by(year, genre) \%>\%
  tally(wt=value) \%>\%
  ungroup \%>\%
  mutate(year=as.Date(sprintf("\%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year") \%>\%
  sg_fill_tableau("purplegray12")
}
}

