% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{sg_header}
\alias{sg_header}
\title{Add labels to the streamgraph}
\usage{
sg_header(sg, header = NULL, as_html = FALSE)
}
\arguments{
\item{sg}{streamgraph object}

\item{header}{Title/header (Note, this is a different implementation of \code{sg_title()}, returning an sg object)}

\item{as_html}{Whether to interpret \code{title}, \code{xaxis_label}, and \code{yaxis_label} arguments as HTML code. Default is FALSE}
}
\value{
sg object
}
\description{
Add labels to the streamgraph
}
\details{
If \code{as_html = TRUE}, the given header will be passed to htmltools::HTML and interpreted as HTML code.
The default (\code{FALSE}) passes the header string as above but as bold, arial, paragraph text.
}
