% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legend.r
\name{sg_legend}
\alias{sg_legend}
\title{Modify streamgraph legend properties}
\usage{
sg_legend(sg, show = FALSE, label = "")
}
\arguments{
\item{show}{if this is \code{TRUE} and \code{interactive} is \code{TRUE} then a popup menu
will be available that lists ll the keys in the data set. Selecting a key will
perform the same action as hovering over the area with the mouse.}

\item{label}{label for the legend (optional)}
}
\description{
If the \code{streamgraph} is interactive, a "legend" can be added
that displays a select menu of all the stream categories. Selecting
a category will highlight that stream in the graph.
}
\details{
TODO: legends for non-interactive streamgraphs
}
\examples{
\dontrun{
library(dplyr)
library(streamgraph)
ggplot2::movies \%>\%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) \%>\%
  tidyr::gather(genre, value, -year) \%>\%
  group_by(year, genre) \%>\%
  tally(wt=value) \%>\%
  ungroup \%>\%
  mutate(year=as.Date(sprintf("\%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year") \%>\%
  sg_fill_brewer("PuOr") \%>\%
  sg_legend(TRUE, "Genre: ")
}
}

