% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{sg_title}
\alias{sg_title}
\title{Add a title to the streamgraph}
\usage{
sg_title(sg, title = "")
}
\arguments{
\item{sg}{streamgraph object}

\item{title}{title}
}
\value{
THIS DOES NOT RETURN AN \code{htmlwidget}!! It returns a \code{shiny.tag}
        class HTML (the widget is wrapped in a \code{<div>}). It should be the LAST
        call in a magrittr pipe chain or called to wrap a streamgraph object for
        output
}
\description{
Add a title to the streamgraph
}
