% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/streamgraph-package.R
\docType{package}
\name{streamgraph-package}
\alias{streamgraph-package}
\title{A package to make \href{http://www.leebyron.com/else/streamgraph/}{streamgraphs}}
\description{
A package to make \href{http://www.leebyron.com/else/streamgraph/}{streamgraphs}
}
\author{
Bob Rudis (@hrbrmstr)
}

