% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{sg_add_marker}
\alias{sg_add_marker}
\title{Add a vertical marker (with optional label) to streamgraph}
\usage{
sg_add_marker(sg, x, stroke_width = 0.5, stroke = "#7f7f7f", space = 5,
  y = 0, label = "", color = "#7f7f7f", size = 12)
}
\description{
Add a vertical marker (with optional label) to streamgraph
}

