% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{sg_annotate}
\alias{sg_annotate}
\title{Add text annotation to streamgraph}
\usage{
sg_annotate(sg, label, x, y, color = "black", size = "12")
}
\description{
Add text annotation to streamgraph
}

