% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{sg_axis_x}
\alias{sg_axis_x}
\title{Modify streamgraph x axis formatting}
\usage{
sg_axis_x(sg, tick_interval = NULL, tick_units = NULL, tick_format = NULL)
}
\arguments{
\item{sg}{streamgraph object}

\item{tick_units}{unit the ticks are in; d3 time scale unit specifier
(defaults to \code{month} for \code{date} scale otherwise not used)}

\item{tick_format}{how to show the labels (subset of \code{strftime}
formatters for \code{date} scale, otherwise \code{sprintf} formats for
\code{continuous} scale) (defaults to \code{\%b} - must specify if \code{continuous}).
See \href{D3 formatting}{https://github.com/mbostock/d3/wiki/Formatting} for more details.}

\item{ticks}{when \code{scale} is \code{date}, \code{ticks} is the interval
between ticks. when \code{scale} is \code{continuous}, \code{ticks} is
the total number of ticks (i.e. "tick count") [defaults to \code{10}]}
}
\value{
streamgraph object
}
\description{
Change the tick interval, units and label text display format for the
streamgraph x axis.
}
\examples{
\dontrun{
library(dplyr)
library(streamgraph)
ggplot2::movies \%>\%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) \%>\%
  tidyr::gather(genre, value, -year) \%>\%
  group_by(year, genre) \%>\%
  tally(wt=value) \%>\%
  ungroup \%>\%
  mutate(year=as.Date(sprintf("\%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year") \%>\%
  sg_axis_x(20, "year", "\%Y")
}
}

