% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{sg_axis_y}
\alias{sg_axis_y}
\title{Modify streamgraph y axis formatting}
\usage{
sg_axis_y(sg, tick_count = 5, tick_format = ",g")
}
\arguments{
\item{sg}{streamgraph object}

\item{tick_count}{number of y axis ticks, not tick interval (defaults to \code{5});
make this \code{0} if you want to hide the y axis}

\item{tick_format}{d3 \href{https://github.com/mbostock/d3/wiki/Formatting#d3_format}{tick format} string}
}
\value{
streamgraph object
}
\description{
Change the tick count
}
\examples{
\dontrun{
library(dplyr)
library(streamgraph)
ggplot2::movies \%>\%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) \%>\%
  tidyr::gather(genre, value, -year) \%>\%
  group_by(year, genre) \%>\%
  tally(wt=value) \%>\%
  ungroup \%>\%
  mutate(year=as.Date(sprintf("\%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year") \%>\%
  sg_axis_x(20, "year", "\%Y") \%>\%
  sg_axis_y(0)
}#'
}

