% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{sg_colors}
\alias{sg_colors}
\title{Modify streamgraph colors}
\usage{
sg_colors(sg, palette = "Blues", text = "black", tooltip = "black")
}
\arguments{
\item{sg}{streamgraph object}

\item{palette}{ColorBrewer pallete atomic character value (defaults to \code{Blues})}

\item{text}{text color CURRENTLY NOT IMPLEMENTED}

\item{tooltip}{color CURRENTLY NOT IMPLEMENTED}
}
\value{
streamgraph object
}
\description{
Change the ColorBrewer palette being used
}
\examples{
\dontrun{
library(dplyr)
library(streamgraph)
ggplot2::movies \%>\%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) \%>\%
  tidyr::gather(genre, value, -year) \%>\%
  group_by(year, genre) \%>\%
  tally(wt=value) \%>\%
  ungroup \%>\%
  mutate(year=as.Date(sprintf("\%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year") \%>\%
  sg_colors("PuOr")
}
}

