% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/streamgraph-package.R
\name{streamgraph-exports}
\alias{\%>\%}
\alias{streamgraph-exports}
\title{streamgraph exported operators}
\description{
The following functions are imported and then re-exported
from the metricsgraphics package to enable use of the magrittr
pipe operator with no additional library calls
}

