% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{streamgraph}
\alias{streamgraph}
\title{Create a new streamgraph}
\usage{
streamgraph(data, key, value, date, width = NULL, height = NULL,
  offset = "silhouette", interpolate = "cardinal", interactive = TRUE,
  scale = "date", top = 20, right = 40, bottom = 30, left = 50)
}
\arguments{
\item{data}{data frame}

\item{key}{bare or quoted name of the category column (defaults to \code{key})}

\item{value}{bare or quoted name of the value column (defaults to \code{value})}

\item{date}{bare or quoted name of the date column (defaults to \code{date})}

\item{width}{Width in pixels (optional, defaults to automatic sizing)}

\item{height}{Height in pixels (optional, defaults to automatic sizing)}

\item{offset}{see d3's \href{https://github.com/mbostock/d3/wiki/Stack-Layout#offset}{offset layout} for more details.
The default is probably fine for most uses but can be one of \code{silhouette} (default),
\code{wiggle}, \code{expand} or \code{zero}}

\item{interpolate}{see d3's \href{https://github.com/mbostock/d3/wiki/SVG-Shapes#area_interpolate}{area interpolation} for more details.
The default is probably fine fore most uses, but can be one of \code{cardinal} (default),
\code{linear}, \code{step}, \code{step-before}, \code{step-after}, \code{basis}, \code{basis-open},
\code{cardinal-open}, \code{monotone}}

\item{interactive}{set to \code{FALSE} if you do not want an interactive streamgraph}

\item{scale}{axis scale (\code{date} [default] or \code{continuous})}

\item{top}{top margin (default should be fine, this allows for fine-tuning plot space)}

\item{right}{right margin (default should be fine, this allows for fine-tuning plot space)}

\item{bottom}{bottom margin (default should be fine, this allows for fine-tuning plot space)}

\item{left}{left margin (default should be fine, this allows for fine-tuning plot space)}
}
\value{
streamgraph object
}
\description{
\code{streamgraph()} initializes the streamgraph html widget
and takes a data frame in "long" format with columns for the
category (by default, it looks for \code{key}) and its associated
\code{date}  and \code{value}. You can supply the names for those
columns if they aren't named as such in your data.\cr
\cr
By default, interactivity is on, but you can disable that by setting
the \code{interactve} parameter to \code{FALSE}.
}
\examples{
\dontrun{
library(dplyr)
library(streamgraph)
ggplot2::movies \%>\%
select(year, Action, Animation, Comedy, Drama, Documentary, Romance, Short) \%>\%
  tidyr::gather(genre, value, -year) \%>\%
  group_by(year, genre) \%>\%
  tally(wt=value) \%>\%
  ungroup \%>\%
  mutate(year=as.Date(sprintf("\%d-01-01", year))) -> dat

streamgraph(dat, "genre", "n", "year")
}
}

