% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/streamgraph.R
\name{streamgraphOutput}
\alias{streamgraphOutput}
\title{Widget output function for use in Shiny}
\usage{
streamgraphOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

